requirements/base.txt
# Bleeding edge Django
django==1.7.1

# Configuration
django-configurations==0.8
django-secure==1.0.1
django-cache-url==0.8.0
dj-database-url==0.3.0

# Forms
django-braces==1.4.0
django-crispy-forms==1.4.0
django-floppyforms==1.2.0

# Models
django-model-utils==2.2

# images
Pillow==2.6.1

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.18.0

# For the persistance stores
psycopg2==2.5.4

# Unicode slugification
unicode-slugify==0.1.1
django-autoslug==1.7.2

# Useful things
django-avatar==2.0

# Time zones support
pytz==2014.10

# Django REST framework
Markdown==2.5.2
django-filter==0.9.1
djangorestframework==3.0.0

# Django extensions
django-extensions==1.4.8

# Django-shorturls (for encoding)
django-shorturls==1.1.0

# AngularJS
django-angular==0.7.10

# Permissions
django-permission==0.8.3

# Revision number
dealer==2.0.4